Fix video player callback not being re-registered
Commit 3bb74f998440179952dc11f80c0521527eaa594f fixed the issue of the
listeners for the next and previous buttons not being re-registered upon
Activity restart in the videos module. However, the media player state
change callback is still not being re-registered in the 'recent videos'
module, causing database updates to not be performed after activity
restarts. This commit fixes the oversight by registering it as well in
the onPostCreate() callback for both tabs (it was previously being done
in the onResume() callback of VideoListActivity i.e. the 'all videos'
tab).