Allow state loss in CourseOutlineFragment addition
CourseOutlineActivity adds CourseOutlineFragment to it's layout after
loading the course components. Since this is done asynchronously, the
exact state is not certain at that time (since the callback method
invocation only guarantees that the Activity has not yet been
destroyed). This can cause the Fragment state to be lost if the state
has already been saved at this time (which can be done at any point),
at which point the FragmentManager will throw an IllegalStateException
to complain about this issue. This commit changes the transaction to
allow state loss to avoid the exception.