Implement tab swiping in videos module
MyVideosTabActivity has had it's FragmentTabHost implementation replaced
with a TabLayout and ViewPager, which allows swiping between the tabs.
This requires the inclusion of the design support library, which is
where TabLayout is implemented.

A global style has been created for TabLayout to match the previous
style applied to the TabWidget (which has been removed). As TabLayout
doesn't define a global theme attribute for it's default style, this
currently needs to be applied individually on each TabLayout definition.
If this becomes a concern at some point, then it could be addressed by
implementing a subclass that defines a default style attribute.

StaticFragmentPagerAdapter has been defined to serve as a standalone
implementation of FragmentPagerAdapter that can be initialized with a
varargs of the data required to set up the Fragment pages. This provides
a simple way to set up a PagerAdapter that represents static Fragment
tabs, and is used for this purpose in MyVideosTabActivity.

Since the landscape mode in MyVideosTabActivity is only for displaying
fullscreen video in MyRecentVideosFragment, the TabLayout is not present
in it's layout and the ViewPager has swiping disabled. This is done by
using DisableableViewPager in that layout, which has also been modified
to use the base enabled flag as the basis for disabling swipe (it also
now supports the android:enabled attribute when inflating from XML,
which has been utilized here).

Since the tabs can now be swiped, and as such need to retain their view
state, the behaviour of MyRecentVideosFragment has been modified to not
remove the PlayerFragment upon being switched out. Since
FragmentPagerAdapter ensures that setUserVisibleHint() is always called
for it's Fragments at the correct times, the explicit invocations of it
in MyRecentVideosFragment have been removed (which had been implemented
in the parent commit to work around the lack of support for it in
FragmentTabHost).