Handle media player from PlayerFragment in ViewPager
PlayerFragment needs to pause the media player when it's not visible,
including when it's page is swiped out in a ViewPager. However, this can
only be inferred by the setUserVisibleHint() call (which is done by the
PagerAdapter implementations for Fragments), and this call is not passed
automatically through to the nested child Fragments. Since
CourseUnitVideoFragment is used inside a ViewPager, and it uses
PlayerFragment as a nested child, a special mode for ViewPager was added
in PlayerFragment in which the media player was controlled by the parent
in commit 2c40de755be0e2c86a47ed92cc65e2d98111d2e3.

This logic is now refactored inside PlayerFragment with the semantically
correct implementation of setUserVisibleHint(), with the assumption that
parent Fragments will explicitly call through to it, in order to have
the logic centralized where it belongs and avoid potential redundancy if
this pattern is reused.