Update screen immediately on new response or comment
- Previously responses disappeared upon adding a new one, because of the
incorrect calls to InfiniteScrollUtils functions, which is fixed in this
commit.

- Also, due to new changes to responses API, we don't get the comments
inside the children array of a response. We instead get a child_count,
depicting the number of comments on a response, which is being used now.

Ultimately fixing the responses screen to update itself properly upon
addition of a new comment as well.