Show pagination spinner for discussion responses
This was accidentally hidden as a byproduct of the changes in commit
22c9137d6c8965f789753368508ca02681d068fc, as it introduced a generic
layout to be referenced for indeterminate ProgressBar implementations,
which was not visible by default, as that is the usual expectation.
However, this caused the progress spinner footer layout for list views
to not show anything by default.

This impacted the pagination on RecyclerView implementations in
InfiniteScrollUtils, which delegates to the adapter to add the row.
ListView implementations were not affected, since the
InfiniteScrollUtils implementation of it just adds a footer view and
controls it's visibility itself. Thus this impacts
CourseDiscussionResponsesFragment, which uses a RecyclerView, but not
CourseDiscussionPostsThreadFragment, which uses a ListView.

This is now fixed in this commit, by explicitly setting the included
layout to visible in the progress spinner footer layout.