Update the comments module to new paginated API
Previously, the whole comment tree was sent at once in the responses
API. The new API unbundles this into separate paginated queries at each
level. Commit 51bd52bced9a0ebb859209f20bb9df35084487b4 added support for
it on showing the comments count in the responses module. This commit
now changes the comments module to use the new API as well, and removes
the unsupported field from the previous API.

The comments adapter has also been fixed to refresh the response header
item as well upon the insertion of a new comment, as it displays the
total count in there.

The responses module has also been updated slightly to match the
comments module, namely cancelling the query tasks upon Fragment
destruction, and modifying the adapter change notification to be more
specific. There is also a small fix on the wrong insertion index being
reported by the comments adapter in the insertCommentAtEnd() method.
Also, the comments adapter has the appropriate nullability annotations
added to match the responses adapter, and the latter had a redundant
null check removed in getItemCount(). The now inaccurate row type label
of MORE_BUTTON is now changed to PROGRESS (as the more button was
changed to a simple autoloading mechanism in commit
53768464022961310bcda7b973086ff38fd4d482).