Remove limit from endorsed responses on threads
Commit 46705eac92463069e1a49e5e14038730b52b0ebc added support for
displaying endorsed responses, which are loaded concurrently with the
non-endorsed ones (but shown before or at the same time as them), but
since it used the same Retrofit method definition for both of them, it
established a page limit to the endorsed responses as well, even though
it was not paginated or meant to be so (as we don't expect a large
number of endorsed responses).

This issue is now fixed by creating a separate method definition for
querying endorsed responses that doesn't include the pagination
parameters, and thus returns all the endorsed responses without
pagination. Comments have been added before each of these methods to
match the discussion responses query method, and that one has also been
updated to reflect the new API format (in which the nested comments are
not included recursively).