Fix wrong thread response count display
Previously, the total comments count were being displayed as the
response count in CourseDiscussionResponsesFragment, as implemented in
commit 7821e17e4665c1e8d20f18978eefbe89f688289c. This included top-level
responses and nested comments. The top-level response count is not
provided in the thread list query API, although it appears upon querying
the details of the individual threads.

Instead of performing yet another query simply to fill in this missing
field, the top-level responses count is now calculated by adding the
count fields provided in the pagination responses for the endorsed and
non-endorsed response query, and the view is hidden from display until
we get both of these responses. The result is then set on the existing
DiscussionThread model (which has an unset state for this field on which
to hide the response count display). Note that since the endorsed
responses are displayed as soon as they're delivered, they may be shown
while the total response count is still not available (and thus not
shown).

The response count layout has also been flattened into a single TextView
to enable easily changing it's visibility directly, which requires the
IconDrawable to be set as a compound drawable dynamically. This is done
by the it's view holder upon inflation to set the default icon, and upon
binding the responses row to choose between the locked and unlocked
items.

Since the pagination loader in CourseDiscussionResponsesFragment is
dealing with the adapter directly now, it has been changed to implement
the InfiniteListController interface instead of just PageLoadCallback
from InfinityScrollUtils. Since this implementation was the reason for
the inclusion of the onPartialPageLoaded() method in the callback
interface (in commit 7821e17e4665c1e8d20f18978eefbe89f688289c), that
method has now been removed from there. Also, the InfiniteListController
interface has been made useful by adding the loadMore() method in it,
and existing usages of PageLoadController have been replaced by just
referencing the interface instead. This also required fixing wrong
loading semantics in PageLoadController, where the loading state was set
to true by default, and all implementations called it's private
onLoadMore() method upon initializing it to make it true.