Fix courseware video autoplay
Commit e22c3a8917751a131b81532cf9ea3bdf200e66ca refactored the media
player automatic handling based on screen events into PlayerFragment.
However, that caused a regression in CourseUnitVideoFragment causing the
autoplay upon swiping the ViewPager to stop working. This is now fixed
by ensuring that the PlayerFragment is initialized once by
CourseUnitVideoFragment upon first receiving visibility to the user
(whether it's opened directly or swiped to), after which the
PlayerFragment will take care of handling the media player based on it's
own lifecycle.