Add pagination for endorsed responses on threads
Commit 46705eac92463069e1a49e5e14038730b52b0ebc added support for
displaying endorsed responses, which are loaded concurrently with the
non-endorsed ones (but shown before or at the same time as them), but
since it used the same Retrofit method definition for both of them, it
established a page limit to the endorsed responses as well, even though
it was not paginated or meant to be so (as we don't expect a large
number of endorsed responses).

This is now fixed by implementing pagination for endorsed responses as
well, which will be completed before starting pagination for
non-endorsed responses. In order to allow multiple paginations to be
performed on the same controller, the PageLoadCallback interface in
InfiniteScrollUtils was changed to an abstract class that defined
another callback method which accepted just the list of results and
assumed that the pagination is continuing, supplementing the original
callback method that took the Page object directly and derived the
pagination state from it. These two methods now call to an abstract
method that takes the pagination state as an explicit parameter.