Fix wrong thread response count display
Previously, the total comments count were being displayed as the
response count in CourseDiscussionResponsesFragment, as implemented in
commit 7821e17e4665c1e8d20f18978eefbe89f688289c. This included top-level
responses and nested comments. The top-level response count is not
provided in the thread list query API, although it appears upon querying
the details of the individual threads.

This is now fixed by querying the thread details explicitly, and
replacing the new model into the adapter. The paginated query for the
responses is intialized concurrently, but their delivery to the adapter
is deferred until the thread details have also been loaded and
delivered.

The response count field has been added to the thread model, with a
default unset state, and the adapter has been set up to hide it's
display in the unset state. The response count layout has also been
flattened into a single TextView to enable easily changing it's
visibility directly, which requires the IconDrawable to be set as a
compound drawable dynamically. This is done by it's view holder upon
inflation to set the default icon, and upon binding the responses row to
choose between the locked and unlocked items.