Fix responses and comments count not incrementing
CourseDiscussionResponsesFragment listens to the responses/comments post
event, and increments their count in it's thread model, which is passed
on to the RecyclerView adapter. However, the model in the adapter is not
guaranteed to be the same object, as it can be replaced with the updated
response from the server after performing PATCH operations on it, such
as voting, following, and flagging. Therefore, it's the current model on
the adapter which needs to be changed, and not the one in the Fragment,
which is not used for any other purpose than to be passed on to the
adapter. This change is implemented by this commit.