Non-updating creation time for discussion responses/comments
Previously whenever notifyDatasetChanged was called on the responses/
comments adapter, the time spanned since the creation time of a response
or a comment updated with respect to latest time.
This is now solved by initialising a time stamp at the time of adapter
creation which is then used as the elapsed time since the creation time
of a response/comment.
The timestamp is then only updated upon addition of new comment/response.