Fix Task implementations
Previously, the Task implementations were catching and handling all
exceptions in the call() method implementation itself, and returning
null in those cases, in spite of the fact that the base RoboAsyncTask
implementation already has a framework for catching the exceptions and
passing it to the onException() callback. The generic exception handling
method that was being called in those cases was presumably not
implemented in this callback to avoid requiring subclasses to call
through to the super implementation. However, since all exceptions were
captured, this guaranteed that the callback was never invoked in the
first place. The callback implementations were also not usually prepared
to handle null results. This inconsistent design led to a lot of issues
and crashes in the case of failures.

This commit fixes these issues by removing all this exception catching
from the call() implementations, and moving the generic error handling
logic to the onException() callback, and ensuring the all the subclasses
call through to the super implementation (and removing the
redundancies in the subclass implementations). In cases where the
callback override is completely redundant, it has been removed. All the
null checking in the call() implementations have been replaced by
@NonNull annotations in the constructor parameters and the local
variables. Since Retrofit API method calls should never return null, the
null checking on the results of tasks that use those methods have been
removed as well. The custom HttpManager using the Apache client that is
used in the courseware module has no such guarantees though, so the null
checks on their results have been left alone.