Halt pagination on receiving errors
The PageLoadCallback interface (more recently converted to an abstract
class) of InfiniteScrollUtils only defined a callback for a successful
page load, and none to handle the case where it encounters an error
doing so. Therefore, lacking support in the callback, the client
implementations mostly didn't handle the case as well (except for the
comments module, which invoked the method to hide the progress bar on
the adapter directly). This issue is now addressed by adding an error
callback that stops the controller, and invoking it in all error cases.