Fix issue with google play info not coming in
If we instantiate an ``HTTPManager`` directly as we do here, then it
doesn't get injected. Hence, the injected ``context`` member was null.
This change makes the context explicit as a constructor argument, so if
someone constructs an ``HTTPManager`` directly rather than through
roboguice, the compiler will force them do the right thing and pass a
context.

The underlying code is pretty bad - google provides APIs for this, so we
shouldn't have make these queries directly, but fixing that seemed out of
scope.

JIRA: https://openedx.atlassian.net/browse/MA-2200