Don't show error message on silent threads refresh
Commit 688c155ceddf941c6e5dc27f238f19225d8c5bc9 introduced silent
reloading of the discussion threads list upon Activity restart.
However, in spite of it being hidden from the user, the reloading logic
displayed an error message on failure, which is quite confusing for the
user. This commit patches the module to ensure that no error message
would be displayed to the user upon performing silent refreshes.