Use better URL resolution logic for course images.
We should be properly resolving URLs that *may* be relative. This came
up when it was discovered that the course image is a relative URL and a
proposal was made to make it an absolute one. Obvs we can't actually do
that since we have bad clients in the wild, but it seemed like we should
fix this and have a utility for it in case it comes up again.