Fix discussions topic selection spinner style
The spinner style provided in the mockups has a non-overlapping and
aligned dropdown integrated with the spinner view, with a simple border
across both of them. Since the spinner border has rounded corners, it
needs to be implemented as a state-list drawable to convert to non-
rounded corners upon the dropdown being displayed. Unfortunately,
Spinner doesn't define any drawable state for the dropdown display.
StatefulSpinner has been implemented to address this deficiency, which
extends from AppCompatSpinner, and duplicates the dropdown popup display
logic along with providing the drawable state.

Note that while rounded corners have been implemented for the dropdown
popup using the background drawable, it doesn't actually clip anything.
The list items have a transparent background and padding so they won't
be an issue normally, but the selector can be drawn on top of the
rounded corners (thus making them non-rounded). The only way to avoid
this is to reimplement ListPopupWindow to use a clipping layout such as
CardView, and reimplement Spinner to use that implementation as well.
Note that even the backport of CardView does not clip it's children, for
performance reasons.

There is also no way to align the Spinner's dropdown window to it's
bounds, since it aligns the content area excluding the padding, and the
dropdown indicator is part of the padding area for the background border
decoration. AlignedDropdownSpinner has been implemented as a subclass of
StatefulSpinner to address this, and it defines attributes and methods
for setting the indicator as an independent drawable, and takes care of
restricting the content measurement and layout to not overlap over it
(though in a pretty hackish and implementation dependant manner).

The indicator caret triangle icon is created as a vector drawable. The
vector drawable implementation has also been backported via the support
library, and the appcompat library and other support libraries have been
updated to the latest stable version (23.3.0), which has support for it.
Some non-public API and bridges duplicating the functionality of private
classes have been created to resolve the indicator drawable and tint
attributes.

The indicator drawable has been set up as a state-list drawable, that
inverts the caret icon when the dropdown is shown. In Lollipop and
above, there is also a rotation animation defined for this state
transition via the animated-vector drawable. This is restricted to
Lollipop and above because state transition animations have not been
backported for state-list drawables.

The current version of the appcompat library doesn't seem to support
loading vector drawables that are nested inside containers (e.g. the
state-list drawable), the TintContextWrapper implementation from the
first alpha release of version 24 (which does support this) has been
copied over, and modified to work with the current
AppCompatDrawableManager implementation. (Strangely, this support seems
to have been reverted in the second alpha release.) The current
appcompat implementation of TintContextWrapper is defined as
package-private anyway.

An API helper class has been added for the deprecated
ViewTreeObserver#removeOnGlobalLayoutListener() method, since this is
not already present in the appcompat library. It's used in the
StatefulSpinner dropdown popup implementation, as well as the existing
PopupMenuHelper implementation.

Since this seems to be the standard style of dropdowns in the design,
the default style for Spinner has been set as this style in the app
theme. This requires usage of AlignedDropdownSpinner everywhere instead
of Spinner. The PopupListView style has been set as the default style
for all dropdown ListView instances in popup windows, and a separator
has been added for the dropdown Spinner style by creating a separate
theme for it, that changes the default to point to a new style that
inherits from the PopupListView style. This seems the optimal
arrangement, as popup windows without separators appear to be more
commonly used in the app (i.e. by both the PopupMenu implementations,
and the popup Spinner style).