Add null data handling in author attribution text
If the author field is null, then if it's just the post label, then just
the time is shown. Otherwise, for endorsement labels, the text view is
hidden. There is also a null check added for the date field, to prevent
a crash from occurring when accessing it's time in that scenario.
Duplicate whitespace are also removed when the date or author label is
null, since it's placeholders in the string template might be in the
middle.

That visibility changes to the text view in
CourseDiscussionResponsesAdapter have also been changed to be called
before invoking this utility method in DiscussionTextUtils, since it can
now cause it to be hidden.