Update profile pic if changed from anywhere
Previously if the profile pic was updated from the web, the changes
didn't had their affect on the drawer that was displayed on MyCourses
screen.
This PR fixes the issue by sending a broadcast whenever we load a user's
profile pic from any screen.
AccountDataLoadedEvent  is broadcasted whenever we get/update a user's
profile, allowing us to update a user's detail on any screen that might
not be immediately in the foreground, but might come up after some
interactions.
- GetAccountTask made concrete, since we are posting the data on the
global event bus as an AccountDataLoadedEvent.