Remove RecyclerView loading indicator item animation
Commit 42610fe774744434695f76726134df2867f0a73b changed the RecyclerView
adapters in the discussion module to report the exact dataset changes to
the RecyclerView instead of just a generic change notification. As a
consequence of providing this detailed information, when a new page of
items were inserted into the adapter, the RecyclerView started to do a
slide animation to move the loading indicator item at the bottom to make
way for the new items to be inserted. Since this is not a desirable UI,
the default item animator is now being removed from the RecyclerView as
part of it's pagination configuration as performed in
InfiniteScrollUtils.