PATCH thread details instead of GET
Previously we were fetching the thread details using a GET
call and setting the read status of the thread using a POST
call (which gave us back the thread details).
Since the GET call was redundant in this case, so it has
been removed in favor of the PATCH call.