MA-2463 Moved receiver from manifest to init
The network connectivity receiver used for offline mode was in the
manifest which caused the app to be launched constantly in the
background. The reciever called the init method of the app which then
reported extraneous app launces to NewRelic. Since offline mode
is only used when the app is being actively used, the reciever
 did not need to be in the Manifest.