Implement Snackbar notifications for new releases
An OkHttp interceptor has been added to capture the headers notifying of
new versions of the app that have been released on the app stores. These
are broadcasted over the event bus and displayed on
MyCoursesListActivity (the main dashboard Activity) in the form of
Snackbars, that will persist until the user swipes them away.

The Activity is treated as a session, with notifications not being
repeated in a single session, but reset after the session is closed.
This design is so that the notifications are always provided directly by
the server, instead of being persisted on the local device. This will
prevent wrongly sent headers to permanently change the app state to an
incorrect one.

There are three different notifications in descending order of priority.
New notifications will only be shown if they have a higher priority than
any other one that has been shown in the current session.

1) Unsupported app version notifications. This is when we get a 426
(upgrade required) error as a response. This means that the API we're
calling has been removed, and will no longer deliver anything.

2) Deprecation notification. This is when we get notified of a new
version of the app, along with a deadline after which the current
version won't have it's API calls supported any longer. These are
prioritized according to the closeness of their deadlines.

3) New release notification. This is when we get notified of a new
version of the app, without any deadline for support of the current
version.  These are prioritized according to the largest new version
number.

For comparing versions, a custom Version class has been written to
strictly represent the app scheme, which can be defined as 'X.X.X.Y',
where X represents numbers, and Y represents an optional alphanumeric
string. Y is ignored for comparison purposes, since it should always
coincide with a change in the X portion. The Version class simply splits
the string by the period token, takes the first three (or less)
substrings it finds, and converts them into numbers. If any of them
can't be converted into an unsigned integer, then it throws a
ParseException (upon encountering which, the NewVersionBroadcastReceiver
implementation terminates it's operations). A test suite for the Version
class has been written as well.

The Snackbar has an action button for updating the app, which attempts
to open the app's section in an app store, using a list of URIs that are
defined in the config. If the initial app store that was used to install
the app can be resolved through one of the URIs, then it's given
priority, and attempted to launch directly. Otherwise, all the URIs are
attempted in order until one of them resolves to a registered Activity.
If none are found, then the action button is not shown. If the app store
was somehow uninstalled after displaying the update button, then it
results in a toast error message.

Unit tests have been added for NewVersionAvailableEvent, and the
NewVersionBroadcastInterceptor, adding functional, validity, and
comparison assertions for all possible permutations.

The event bus posting verification tests (for
NewVersionBroadcastInterceptor and the static post() method)
unfortunately doesn't work in the parameterized test suite at the
moment, since the event bus initialization requires loading the Android
framework, and we currently don't have a parameterized Robolectric test
runner implementation. Nor can RoboGuice be used to inject a mock
instance of the event bus, since initializing RoboGuice also requires a
Context (which is only available in the Android framework). Therefore,
these tests are being ignored for now.

No tests have been written for verifying the event bus registration at
the UI level, and the subsequent display of the appropriate Snackbar
notification, since there is currently no official shadow implementation
for Snackbar in the Robolectric framework. If a custom Snackbar shadow
is added to the project at some point, then this test can be added as
well at that point.