Add custom exception throwing declaration in LoginService
The Retrofit interface for the login APIs weren't defined to throw the
custom checked exception wrapper (RetroHttpException) that we throw from
the global Retrofit error handler. As a result, upon encountering this
undeclared checked exception thrown by Retrofit's invocation handler,
the framework throws an unchecked UndeclaredThrowableException that
wraps around it instead, which the caller is not prepared to catch. This
has the side-effect of disabling the error handling logic for the login
request. This commit fixes the issue by declaring all the method
definitions in LoginService to throw RetroHttpException.

The OkHttp authenticator for refreshing the OAuth token has also been
updated to catch the RetroHttpException, and if it's wrapping an
IOException then throw that, or otherwise just return null. The
RetroHttpException class and it's various subclasses have also been
updated to provide type-safe methods for retreiving their wrapped
exceptions/errors.

There was an instance of wrapping an IOException manually in an
RetroHttpException wrapper in the Retrofit API accessor for getting user
enrolled courses in UserAPI, which was performing manual caching of the
courses. These have been updated to wrap around custom-built (sparse)
RetrofitError wrappers to conform with the newly added type-safety,
and the custom caching has been deprecated in favor of using OkHttp's
own automatic caching mechanism.