Improve error handling logic for login request
The login task was catching all exceptions itself and returning null,
thus forcing the success callback to check for the validity and
availibity of the data, instead of throwing the exceptions and letting
the exception callback handle them. It didn't catch the IO exceptions
from the profile data request that it performed using the old Apache
HTTP client though, and the exception callback wasn't defined to handle
these either.

This commit consolidates the error handling logic in the exception
callback, and ensures that a RetrofitHttpException will always be thrown
in case of an error, or in the case of the get profile request failing
with a generic error or failing validation, the custom login exception
that was being used originally from the success callback would be
thrown.

The redundant declaration to throw the login exception has also been
removed from the user success callback method in the MobileLoginCallback
interface.