Clear user data upon logout + migrations
- Sha1Util and all its usages refactor
- Previously NoSuchAlgorithmException & UnsupportedEncodingException
were thrown if the encryption failed. But now the function itself
catches these exceptions and sends back the plain text if encryption
failed.
- Migrations and folders restructuring
- Separate folder for videos made in which SHA1 hashed username based
folders are made for each logged in user, storing their videos in it.
- Subtitles folder renamed and placed in the videos folder.
- Tests added for SHA1 util function.
- Nuke shared preferences
- Silently fetch and cache course structure of downloaded videos
For videos to appear correctly i.e. appear in the same order as they
appear in courseware, the course structure needs to be cached.
For example, if a user has downloaded multiple videos and then logs out,
we delete all the cached data. But upon logging back in, the downloaded
videos need to appear correctly which requires the videos' course's
structure.
This commit addresses the case mentioned above by first querying the
course ids from the downloaded videos that exist in our db and then
fetching and caching the course structure of those course upon login.
- Database version definition rephrased.
- Separate tests added for all utility functions in Shat1Util class
- Database's upgrade function turned into a nested transaction.
- App's usage of external storage standardised using the
FileUtil#getAppExternalDir function.
- TranscriptManager refactored.
- Separate function for manual logout on button tap
- Singleton issues fix for CacheManager
Since we are now manually deleting the directories and files in the app
directory and external directory, the cache directory also gets deleted.
CacheManager being a singleton class still thinks that the directory its
using for storing cached files is available causing
FileNotFoundExceptions upon access.