Add a list of constants for HTTP status codes
There wasn't any existing constants list for HTTP status codes that was
comprehensive previously. The codes were either directly referenced as
literals, which was bad for readability and made it easy to make and
retain mistakes, or the constants defined in HttpURLConnection or
HttpStatus (from the deprecated Apache's HTTP client) were being
referenced, which were not comprehensive.

This commit addresses the issue by defining a custom HttpStatus class in
the project that contains a comprehensive list of the status codes.
These were copied from the HttpResponseStatus class in the Netty
framework, and modified to be constant int variables instead of
instances of it's own class.