Invalidate refresh token on logout
Upon logging out, and before clearing the access and refresh tokens,
send a request to the server to revoke the refresh token and all the
access tokens granted on it's basis. This request would only go through
if the device is online at the time of logging out; otherwise the tokens
will just be deleted locally.