Update AssertJ Core from 1.x to 2.x
AssertJ uses a different major version number for each version of Java,
with 1.x using Java 6 specific assertions, 2.x adding Java 7 assertions,
and 3.x Java 8 assertions. Unfortunately, development on the 1.x branch
has been discontinued since 1.7.1, and thus it missed many core
convenience assertions that were added later in the 2.x and 3.x
branches.

Fortunately, as of 2.3.0 and 3.3.0 respectively, the assertions that are
specific to Java 6 have been isolated into a special class named
Java6Assertions to support Android, which doesn't have the full
complement of Java 7 APIs even in KitKat+. Therefore, even though the
Java 7 assertions aren't in use or needed, the AssertJ Core test
dependency is still being updated from version 1.7.1 (the latest 1.x
version) to 2.5.0 (the latest 2.x version), in order to be on a live and
updated branch. All the existing assertions are now updated to import
the Java6Assertions class instead of the generic Assertions class (which
now includes Java 7 APIs).