Profile pictures put behind a feature flag
DISCUSSIONS_ENABLE_PROFILE_PICTURE_PARAM has been introduced to enable/
disable the profile pictures in Discussions.

Currently stage has some obsolete data which doesn't allow profile
pictures to work properly on it. So, this commit the issue by putting
the feature behind a feature-flag.

If the flag is disabled we don't send the query param required for GET
request to send profile pictures in response.