Don't assume TaskMessageCallback from Context
Before upgrading to Retrofit 2 in the PR #799, TaskMessageCallback had
to be explicitly provided to a Task so that message callbacks may be
called by the Task implementation accordingly.

While upgrading to Retrofit 2 this little detail got missed and the
constructors in ErrorHandlingCallback assumed TaskMessageCallback from
the context param being provided. Since our base activity implements
the TaskMessageCallback, so it caused error dialogs to appear in case
of non-user invoked call triggers as well.