Remove reviewers checkboxes from PR template
The checkboxes were previously useful to check the review approvals
status in one place. Now this information can be found in the banner at
the bottom in GitHub's new review UI, which makes the checkboxes
redundant. Accordingly, they are being removed in this commit.
The review instructions have also been updated to instruct reviewers to
use the new review UI.