Update HTTP error exception to contain response
The HttpResponseStatusException was being used as an exception to throw
upon encountering an error status code in an HTTP response. However, it
didn't hold any data other than the status, even though the catcher
might need to examine other response data. This commit addresses the
deficiency by modifying the exception to take the whole Retrofit
response object, which can be queried by the catchers as needed.