Catch exceptions thrown by the database helper
Originally, the DbHelper implementation caught all exceptions from
inside the onUpgrade() callback. This silent failure mode resulted in an
inconsistent state where the only version would be incremented, but not
the actual schema.

This issue was fixed in commit 0cef258b719852c437a1830d3275e64e15233e30,
by removing the general exception catching block from the onUpgrade()
callback, thus allowing the getDatabase() call to throw the exception
instead. However, this was also problematic because the callers weren't
catching any exceptions, which would result in the app crashing upon
database creation or upgrade failures.

This commit fixes this new issue by ensuring that the caller catches any
SQLite exceptions thrown by the getDatabase() calls. The caller in this
case is the IDatabaseBaseImpl#execute() method, which has now been
modified to return the default value of the operation in the case of
encountering this exception, in addition to delivering the exception to
the callback, and logging it in Crashlytics. The exception handling
mechanism in the base database operation implementation
(DbOperationBase#requestExecute()) has also been modified to log the
exceptions in Crashlytics.