Fix generics usage in database query method
The execute() and enqueue() methods in IDatabaseBaseImpl were not using
matching generic types properly to ensure that that operation type
matched the actual return type. Instead, the operation type was defined
as a wildcard. This was actually covering over a few bugs in the
implementation, where there were different types being returned, due to
it missing a few conversion steps. However, these bugs weren't exposed
since they were only there for the synchronous return type, which was
returned as null when it was invoked asynchronously by providing a
callback, and all the usages were providing callbacks.

This commit fixes this issue by ensuring the matching of the generic
types of the operation and the returned value, and fixing the improper
usages in the implementations by adding the missing conversion logic.