Remove error notifications from NavigationFragment
Since the navigation drawer usually isn't being displayed, it doesn’t
make sense to have it show error notifications to the user. Even when
the drawer is open and on display, there's no appropriate area to show a
data loading error that’s associated with it. Therefore, this commit
removes the Snackbar error notification logic from calls made by
NavigationFragment.

Note that the drawer was also added automatically to many Activities
where it wasn’t even used, and was always blocked from opening, and
where cases showing an error was completely inappropriate. This was
because the default layout of the BaseSingleFragmentActivity included
the drawer’s layout, and subclasses that didn’t support it (which were
the majority) blocked it on a case-by-case basis.